requirements.txt
opencv-python
flask
flask-cors
ultralytics
mongoengine
pillow
python-dotenv
werkzeug
transformers
google-api-python-client
google-generativeai
google-genai